+++ resources/release-notes.md
@@ -1,3 +1,7 @@
 ## Software
 
 - (All) Add chart for plotting shaded regions (`mag.graphics.chart.Region`)
+
+## Package
+
+- Update code analyzer configuration to 1.2.0 (allow class name starting with `t` for test cases)